test: extract shared mocks and fakes into testutils_test.go
